# src/main/resources/APIDOC.YAML
openapi: 3.0.1
info:
  title: Constructa OpenApi specification
  description: OpenApi documentation for Constructa BE
  contact:
    name: Ondřej Bureš
    url: 'https://digitalnoodles.cz/'
    email: bureson2@fel.cvut.cz
  version: '1.0'
servers:
  - url: 'http://localhost:8080/'
    description: Local ENV
security:
  - bearerAuth: []
paths:
  /api/v1/work-reports:
    get:
      tags:
        - Work reports
      operationId: getWorkReports
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkReportDTO'
      security:
        - bearerAuth: []
    put:
      tags:
        - Work reports
      operationId: updateWorkReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkReportRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkReportDTO'
      security:
        - bearerAuth: []
    post:
      tags:
        - Work reports
      operationId: createWorkReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkReportRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkReportDTO'
      security:
        - bearerAuth: []
...
  /api/v1/work-reports/illness:
    post:
      tags:
        - Work reports
      operationId: recordIllness
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IllnessRequest'
        required: true
      responses:
        '201':
          description: Created
      security:
        - bearerAuth: []
  /api/v1/work-reports/end-work:
    post:
      tags:
        - Work reports
      operationId: stopRecordAttendance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopAttendanceRequest'
        required: true
      responses:
        '201':
          description: Created
      security:
        - bearerAuth: []
  /api/v1/work-reports/attendance:
    post:
      tags:
        - Work reports
      operationId: recordAttendance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkReportDTO'
      security:
        - bearerAuth: []
...
  /api/v1/auth/register:
    post:
      tags:
        - Authentization
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
  /api/v1/auth/authenticate:
    post:
      tags:
        - Authentization
      operationId: authenticate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
  '/api/v1/work-reports/{workReportId}':
    get:
      tags:
        - Work reports
      operationId: getWorkReport
      parameters:
        - name: workReportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkReportDTO'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Work reports
      operationId: deleteWorkReport
      parameters:
        - name: workReportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
  '/api/v1/work-reports/user/{userId}':
    get:
      tags:
        - Work reports
      operationId: getUsersWorkReports
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkReportDTO'
      security:
        - bearerAuth: []
  /api/v1/work-reports/my:
    get:
      tags:
        - Work reports
      operationId: getMyWorkReports
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkReportDTO'
      security:
        - bearerAuth: []
  '/api/v1/work-reports/location/{locationId}':
    get:
      tags:
        - Work reports
      operationId: getWorkLocation
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDTO'
      security:
        - bearerAuth: []
...
  '/api/v1/vehicles/reports/{vehicleReportId}':
    get:
      tags:
        - Vehicle reports
      operationId: getVehicleReport
      parameters:
        - name: vehicleReportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleReportDTO'
      security:
        - bearerAuth: []
...
  '/api/v1/vehicles/reports/{reportId}':
    delete:
      tags:
        - Vehicle reports
      operationId: deleteVehicleReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
components:
  schemas:
    WorkReportRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        minutes:
          type: integer
          format: int32
        type:
          type: string
          enum:
            - WORK_REPORT
            - HOLIDAY
            - SICK_NOTE
            - MATERNITY_LEAVE
            - UNPAID_VACATION
            - SICK_DAY
        userId:
          type: integer
          format: int64
    LocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        city:
          type: string
        street:
          type: string
        descriptiveNumber:
          type: string
        name:
          type: string
        country:
          type: string
        postCode:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        titleBeforeName:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        titleAfterName:
          type: string
        bankAccount:
          type: string
        dateOfAcceptance:
          type: string
          format: date-time
        dateOfBirth:
          type: string
          format: date-time
        birthId:
          type: string
        hourRate:
          type: integer
          format: int32
        monthSalary:
          type: integer
          format: int32
        address:
          $ref: '#/components/schemas/LocationDTO'
    WorkReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        minutes:
          type: integer
          format: int32
        reportingEmployee:
          $ref: '#/components/schemas/UserDTO'
        location:
          $ref: '#/components/schemas/LocationDTO'
        type:
          type: string
...
    VehicleDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        factory:
          type: string
        name:
          type: string
        registrationNumber:
          type: string
        conditionMotorcycleWatch:
          type: number
          format: double
        mileage:
          type: number
          format: double
        type:
          type: string
        boughtAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        qrCode:
          type: string
        vinCode:
          type: string
...
    VehicleReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        originalConditionMotorcycleWatch:
          type: number
          format: double
        afterworkConditionMotorcycleWatch:
          type: number
          format: double
        cargoMass:
          type: integer
          format: int32
        cargoType:
          type: string
        distance:
          type: number
          format: double
        purchaseOfFuelLitres:
          type: number
          format: double
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        description:
          type: string
        vehicle:
          $ref: '#/components/schemas/VehicleDTO'
        driver:
          $ref: '#/components/schemas/UserDTO'
...
    IllnessRequest:
      type: object
      properties:
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
    StopAttendanceRequest:
      type: object
      properties:
        time:
          type: integer
          format: int32
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
    AttendanceRequest:
      type: object
      properties:
        locationId:
          type: integer
          format: int64
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
    RegisterRequest:
      type: object
      properties:
        titleBeforeName:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        titleAfterName:
          type: string
        bankAccount:
          type: string
        email:
          type: string
        password:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        birthId:
          type: string
        hourRate:
          type: integer
          format: int32
        monthSalary:
          type: integer
          format: int32
        phone:
          type: string
        role:
          type: string
          enum:
            - ROLE_ADMIN
            - ROLE_MANAGER
            - ROLE_REPORTER
            - ROLE_WAREHOUSE_MANAGER
            - ROLE_MECHANICAL_ENGINEER
            - ROLE_CONSTRUCTION_MANAGER
            - ROLE_DRIVER
            - ROLE_WORKER
            - ROLE_EMPLOYEE
            - ROLE_EXTERNALIST
            - ROLE_FIRED
        country:
          type: string
        city:
          type: string
        street:
          type: string
        descriptiveNumber:
          type: string
        postCode:
          type: string
    AuthenticationResponse:
      type: object
      properties:
        token:
          type: string
    AuthenticationRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT auth description
      in: header
      scheme: bearer
      bearerFormat: JWT